Fix insert code block width / selection
Add a wrapper around the code blocks.  This wrapper will take its size
from the code contents, thus being full width _and_ be the immediate
parent of the code elements.  The codehilite block can therefore
overflow and have a limited viewport / size that doesn't matter to the
sizing of the child, as we established that its size will be dictated by
its contents.  CSS is awesome.

Fixes #276 while maintaining the fix for #252